requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
websocket-client>=1.4.0
ta>=0.10.0
PyYAML>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
asyncio-mqtt>=0.13.0
streamlit>=1.25.0
plotly>=5.15.0
